feat: Queen 이동 검증 추가 및 이동 검증 전략 분리
