+++ docs/README.md
@@ -110,11 +110,15 @@ abcdefgh (file)
 - [ ] `target` 위치의 기물이 본인의 기물인 경우 이동 불가능하다.
 
 - [x] ROOK
-  - `target`의 `index`가 `File`이나 `Rank` 둘 중에 하나와 같아야 한다.
+  - `target`이 가로, 세로 방향에 있어야 한다.
   - 경로에 기물이 존재할 경우 이동 불가능하다.
 
 - [x] BISHOP
   - `target`이 대각선 방향에 있어야 한다.
+  - 경로에 기물이 존재할 경우 이동 불가능하다.
+
+- [x] QUEEN
+  - `target`이 가로, 세로, 대각선 방향에 있어야 한다.
   - 경로에 기물이 존재할 경우 이동 불가능하다.
 
 #### Direction
