+++ src/main/java/chess/domain/piece/Bishop.java
@@ -6,6 +6,8 @@ import static chess.domain.position.Direction.TOP_LEFT;
 import static chess.domain.position.Direction.TOP_RIGHT;
 
 import chess.domain.Color;
+import chess.domain.piece.strategy.LinearMovingStrategy;
+import chess.domain.piece.strategy.MovingStrategy;
 import chess.domain.position.Direction;
 import chess.domain.position.Position;
 import java.util.List;
@@ -15,39 +17,15 @@ public class Bishop extends Piece {
     private static final List<Direction> DIRECTIONS = List.of(TOP_LEFT, TOP_RIGHT, BOTTOM_LEFT, BOTTOM_RIGHT);
     private static final String NOTATION = "B";
 
+    private final MovingStrategy movingStrategy;
+
     public Bishop(Color color) {
         super(color);
+        this.movingStrategy = new LinearMovingStrategy(DIRECTIONS);
     }
 
     public void validateMove(List<List<Piece>> board, Position sourcePosition, Position targetPosition) {
-        Direction direction = Direction.of(sourcePosition, targetPosition);
-        validateDirection(direction);
-
-        Position currentPosition = sourcePosition;
-        while (!currentPosition.equals(targetPosition)) {
-            currentPosition = currentPosition.add(direction);
-            Piece currentPiece = findPiece(board, currentPosition);
-            validateExistPiece(currentPiece);
-        }
-    }
-
-    private void validateDirection(Direction direction) {
-        if (!DIRECTIONS.contains(direction)) {
-            throw new IllegalArgumentException("해당 기물이 갈 수 없는 경로입니다.");
-        }
-    }
-
-    private void validateExistPiece(Piece currentPiece) {
-        if (!currentPiece.isEmpty()) {
-            throw new IllegalArgumentException("경로에 기물이 존재하여 이동할 수 없습니다.");
-        }
-    }
-
-    private Piece findPiece(List<List<Piece>> board, Position sourcePosition) {
-        int rankIndex = sourcePosition.getRankIndex();
-        int fileIndex = sourcePosition.getFileIndex();
-
-        return board.get(rankIndex).get(fileIndex);
+        movingStrategy.validateMove(board, sourcePosition, targetPosition);
     }
 
     @Override
