+++ src/main/java/chess/domain/piece/strategy/MovingStrategy.java
@@ -1,0 +1,11 @@
+package chess.domain.piece.strategy;
+
+import chess.domain.piece.Piece;
+import chess.domain.position.Position;
+import java.util.List;
+
+@FunctionalInterface
+public interface MovingStrategy {
+
+    void validateMove(List<List<Piece>> board, Position sourcePosition, Position targetPosition);
+}
